optimization of methods done

+++ services/implementation/RouteServiceImpl.java
@@ -10,7 +10,7 @@ import java.util.List;
 
 public class RouteServiceImpl implements RouteService {
 
-    private RouteRepo routeRepo;
+    private final RouteRepo routeRepo;
 
     public RouteServiceImpl(RouteRepo routeRepo) {
         this.routeRepo = routeRepo;
@@ -18,16 +18,35 @@ public class RouteServiceImpl implements RouteService {
 
     @Override
     public Route addRoute(Route route) {
+        if (route != null) {
+            return routeRepo.add(route);
+        } else {
+            System.err.println("Введено некоректное значение ! ");
+        }
         return routeRepo.add(route);
     }
 
     @Override
     public void removeRoute(Integer id) {
-        routeRepo.remove(id);
+        if (routeRepo.findById(id) != null) {
+            if (routeRepo.findById(id).getTransport() == null) {
+                routeRepo.remove(id);
+                System.out.println("Маршрут успешно удален !");
+            } else {
+                System.err.println("Невозможно удалить маршрут ! Данный маршрут зареплен за транспортом !");
+            }
+        } else {
+            System.err.println("Маршрута с таким id не существует !");
+        }
     }
 
     @Override
     public Route findRouteById(Integer id) {
+        if (routeRepo.findById(id) != null) {
+            return routeRepo.findById(id);
+        } else {
+            System.err.println("Маршрута с данным id не существует");
+        }
         return routeRepo.findById(id);
     }
 
@@ -46,6 +65,11 @@ public class RouteServiceImpl implements RouteService {
                 temp.add(r);
             }
         }
+
+        if (temp.isEmpty()) {
+            System.err.println("Лист пуст !");
+        }
+
         return temp;
     }
 }
